round to roundf

+++ R2+1D/Conv3d.cpp
@@ -44,7 +44,7 @@ void Conv3d(dtype* X_data, int_t* X_num, dtype* Y_data, int_t* Y_num, ktype* Ker
 											if(dPos >= 0 && hPos >= 0 && wPos >= 0 && dPos < XD && hPos < XH && wPos < XW)
 												tmp_Y += ((int_t)X_data[xn*XC*XD*XH*XW + xc*XD*XH*XW + dPos*XH*XW + hPos*XW + wPos]- zp_in) * Kernel_data[yc*XC*KD*KH*KW + xc*KD*KH*KW + kd*KH*KW + kh*KW + kw];
 										}
-						tmp_Y = round((ftype)tmp_Y*(scale_in*kernel_scale[yc]/scale_out) + zp_out);
+						tmp_Y = roundf((ftype)tmp_Y*(scale_in*kernel_scale[yc]/scale_out) + zp_out);
             			Y_data[yPos] = (tmp_Y > 255) ? 255 : (tmp_Y < 0) ? 0 : (dtype)tmp_Y;
 					}
 	return;
